Console-Like Output And Small Tweaks
Input for the program is slightly more customizable. Coders can insert
references to custom processes to be executed, which can then be run. I
used this to make the trigonometric functions optional, in case a user
doesn't want to load one of the packages. The function memory can also
be wiped with a single command.